move configs around
