+++ package.json
@@ -17,9 +17,9 @@
 		"clean:dist": "rm -rf dist/*",
 		"clean:test-page": "rm -rf .test-page/*",
 		"compile": "rollup --config",
-		"dev": "yarn clean:test-page && rollup --config rollup.config.test-page.js --watch",
+		"dev": "yarn clean && yarn compile --watch",
 		"fix": "yarn lint --fix",
-		"gh-pages:build": "yarn clean:test-page && rollup --config rollup.config.test-page.js",
+		"gh-pages:build": "yarn build",
 		"gh-pages:serve": "yarn gh-pages:build && serve .test-page",
 		"lint": "eslint . --ext=js,ts,tsx",
 		"prepublishOnly": "yarn build",
