+++ rollup.config.js
@@ -1,6 +1,6 @@
 /* eslint-disable import/no-default-export */
 
-import packageConfig from './rollup.config.package';
-import testPageConfig from './rollup.config.test-page';
+import packageConfig from './src/rollup.config';
+import testPageConfig from './test-page/rollup.config';
 
 export default [packageConfig, testPageConfig];
